requirements.txt
python-dotenv
PyYAML
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pandas
rich
rich.logging
matplotlib
openpyxl
python-docx
python-dateutil
click